requirements.txt
GeoAlchemy2==0.17.1
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
Pygments==2.19.1
SQLAlchemy==2.0.41
accelerate==1.7.0
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
bitsandbytes==0.46.0
certifi==2025.4.26
charset-normalizer==3.4.2
click-plugins==1.1.1
click==8.2.1
cligj==0.7.2
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.6.0
diffusers==0.33.1
dill==0.3.8
docstring_parser==0.16
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.0
geopandas==1.1.0
greenlet==3.2.3
hf-xet==1.1.3
hf_transfer==0.1.9
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
mercantile==1.2.1
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
peft==0.15.2
pillow==11.2.1
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.9.0
shapely==2.1.1
shtab==1.7.2
six==1.17.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
triton==3.3.0
trl==0.18.1
typeguard==4.4.3
typing_extensions==4.14.0
tyro==0.9.24
tzdata==2025.2
ultralytics-thop==2.0.14
ultralytics==8.3.154
unsloth==2025.6.2
unsloth_zoo==2025.6.1
urllib3==2.4.0
xformers==0.0.30
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
Flask==3.0.3
